--- packages/ui-design-tool/components/AxisGrid/types.ts
@@ -1,8 +1,0 @@
-import { UIRecordKey } from '@/types/Identifier';
-import ReactTypes from '@pigyuma/react-utility-types';
-
-export type AxisGridProps = ReactTypes.UnknownProps;
-
-export type AxisGridRendererProps = {
-  recordKey?: UIRecordKey;
-};
